requirements.txt
# ========================================
# CORE ML & RL
# ========================================
torch>=2.0.0
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0
gymnasium>=0.29.0

# ========================================
# DATA & TECHNICAL ANALYSIS
# ========================================
yfinance>=0.2.33
pandas>=2.1.4
numpy>=1.26.2
ta>=0.11.0
scipy>=1.11.0  # Pour drift detection (PSI, KS Test)

# ========================================
# TRADING
# ========================================
alpaca-py>=0.25.0

# ========================================
# MONITORING
# ========================================
wandb>=0.15.0
tensorboard>=2.13.0
prometheus-client>=0.19.0

# ========================================
# DATABASE
# ========================================
psycopg2-binary>=2.9.9

# ========================================
# WEB
# ========================================
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.25.0
plotly>=5.18.0

# ========================================
# ADVISORY ENGINE
# ========================================
finnhub-python>=2.4.0
vaderSentiment>=3.3.2
statsforecast>=1.7.0
ollama>=0.3.0
cachetools>=5.3.0

# ========================================
# UTILS
# ========================================
python-dotenv>=1.0.0
requests>=2.31.0
